module_3/requirements.txt
psycopg2-binary
python-dotenv
flask